+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-07-11
+
+### Added
+
+- Ability to remove minutes (eg. 8AM)
+- Ability to remove leading 0 padding from hours
+
+### Changed
+
+- Define prop types and defaults
+
 ## [1.1.2] - 2023-07-07
 
 ### Fixed
@@ -66,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add "Inspired by" on README
 
-[unreleased]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.1.2...HEAD
+[unreleased]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/markandrewkato/vue3-time-picker-plugin/compare/v1.0.12...v1.1.0